Network.AWS.Greengrass.AssociateServiceRoleToAccount

Description

Associates a role which is used by AWS Greengrass. AWS Greengrass uses the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. It needs to have minimum permissions in policy `AWSGreengrassResourceAccessRolePolicy`

Creating a Request

associateServiceRoleToAccount :: AssociateServiceRoleToAccount

Creates a value of AssociateServiceRoleToAccount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asrtaRoleARN - Role arn you wish to associate with this account.

data AssociateServiceRoleToAccount

See: associateServiceRoleToAccount smart constructor.

Request Lenses

asrtaRoleARN :: Lens' AssociateServiceRoleToAccount (Maybe Text)

Role arn you wish to associate with this account.

Destructuring the Response

associateServiceRoleToAccountResponse

Arguments

:: Int
asrtarsResponseStatus

-> AssociateServiceRoleToAccountResponse

Creates a value of AssociateServiceRoleToAccountResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asrtarsAssociatedAt - Time when the service role was associated to the account.

asrtarsResponseStatus - -- | The response status code.

data AssociateServiceRoleToAccountResponse

See: associateServiceRoleToAccountResponse smart constructor.

Response Lenses

asrtarsAssociatedAt :: Lens' AssociateServiceRoleToAccountResponse (Maybe Text)

Time when the service role was associated to the account.

asrtarsResponseStatus :: Lens' AssociateServiceRoleToAccountResponse Int

- | The response status code.