Network.AWS.Greengrass.CreateCoreDefinition

Description

Creates a core definition. You may optionally provide the initial version of the core definition or use 'CreateCoreDefinitionVersion' at a later time. AWS Greengrass Groups must each contain exactly 1 AWS Greengrass Core.

Creating a Request

createCoreDefinition :: CreateCoreDefinition

Creates a value of CreateCoreDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccdAmznClientToken - The client token used to request idempotent operations.

ccdInitialVersion - Information on the initial version

ccdName - name of the core definition

data CreateCoreDefinition

Information on the core definition request

See: createCoreDefinition smart constructor.

Request Lenses

ccdAmznClientToken :: Lens' CreateCoreDefinition (Maybe Text)

The client token used to request idempotent operations.

ccdInitialVersion :: Lens' CreateCoreDefinition (Maybe CoreDefinitionVersion)

Information on the initial version

ccdName :: Lens' CreateCoreDefinition (Maybe Text)

name of the core definition

Destructuring the Response

createCoreDefinitionResponse

Arguments

:: Int
ccdrsResponseStatus

-> CreateCoreDefinitionResponse

Creates a value of CreateCoreDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccdrsLatestVersionARN - Latest version arn of the definition.

ccdrsARN - Arn of the definition.

ccdrsName - Name of the definition.

ccdrsCreationTimestamp - Timestamp of when the definition was created.

ccdrsId - Id of the definition.

ccdrsLatestVersion - Last version of the definition.

ccdrsLastUpdatedTimestamp - Last updated timestamp of the definition.

ccdrsResponseStatus - -- | The response status code.

data CreateCoreDefinitionResponse

See: createCoreDefinitionResponse smart constructor.

Response Lenses

ccdrsLatestVersionARN :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Latest version arn of the definition.

ccdrsARN :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Arn of the definition.

ccdrsName :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Name of the definition.

ccdrsCreationTimestamp :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Timestamp of when the definition was created.

ccdrsId :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Id of the definition.

ccdrsLatestVersion :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Last version of the definition.

ccdrsLastUpdatedTimestamp :: Lens' CreateCoreDefinitionResponse (Maybe Text)

Last updated timestamp of the definition.

ccdrsResponseStatus :: Lens' CreateCoreDefinitionResponse Int

- | The response status code.