Network.AWS.Greengrass.CreateDeviceDefinitionVersion

Description

Creates a version of a device definition that has already been defined.

Creating a Request

createDeviceDefinitionVersion

Arguments

:: Text
cddvDeviceDefinitionId

-> CreateDeviceDefinitionVersion

Creates a value of CreateDeviceDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cddvAmznClientToken - The client token used to request idempotent operations.

cddvDevices - Devices in the definition version.

cddvDeviceDefinitionId - device definition Id

data CreateDeviceDefinitionVersion

See: createDeviceDefinitionVersion smart constructor.

Request Lenses

cddvAmznClientToken :: Lens' CreateDeviceDefinitionVersion (Maybe Text)

The client token used to request idempotent operations.

cddvDevices :: Lens' CreateDeviceDefinitionVersion [Device]

Devices in the definition version.

cddvDeviceDefinitionId :: Lens' CreateDeviceDefinitionVersion Text

device definition Id

Destructuring the Response

createDeviceDefinitionVersionResponse

Arguments

:: Int
cddvrsResponseStatus

-> CreateDeviceDefinitionVersionResponse

Creates a value of CreateDeviceDefinitionVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cddvrsARN - Arn of the version.

cddvrsCreationTimestamp - Timestamp of when the version was created.

cddvrsVersion - Unique Id of a version.

cddvrsId - Id of the resource container.

cddvrsResponseStatus - -- | The response status code.

data CreateDeviceDefinitionVersionResponse

See: createDeviceDefinitionVersionResponse smart constructor.

Response Lenses

cddvrsARN :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

Arn of the version.

cddvrsCreationTimestamp :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

Timestamp of when the version was created.

cddvrsVersion :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

Unique Id of a version.

cddvrsId :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

Id of the resource container.

cddvrsResponseStatus :: Lens' CreateDeviceDefinitionVersionResponse Int

- | The response status code.