Network.AWS.Greengrass.CreateGroup

Description

Creates a group. You may optionally provide the initial version of the group or use 'CreateGroupVersion' at a later time.

Creating a Request

createGroup :: CreateGroup

Creates a value of CreateGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgAmznClientToken - The client token used to request idempotent operations.

cgInitialVersion - Information on the initial version

cgName - name of the group

data CreateGroup

See: createGroup smart constructor.

Request Lenses

cgAmznClientToken :: Lens' CreateGroup (Maybe Text)

The client token used to request idempotent operations.

cgInitialVersion :: Lens' CreateGroup (Maybe GroupVersion)

Information on the initial version

cgName :: Lens' CreateGroup (Maybe Text)

name of the group

Destructuring the Response

createGroupResponse

Arguments

:: Int
cgrsResponseStatus

-> CreateGroupResponse

Creates a value of CreateGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgrsLatestVersionARN - Latest version arn of the definition.

cgrsARN - Arn of the definition.

cgrsName - Name of the definition.

cgrsCreationTimestamp - Timestamp of when the definition was created.

cgrsId - Id of the definition.

cgrsLatestVersion - Last version of the definition.

cgrsLastUpdatedTimestamp - Last updated timestamp of the definition.

cgrsResponseStatus - -- | The response status code.

data CreateGroupResponse

See: createGroupResponse smart constructor.

Response Lenses

cgrsLatestVersionARN :: Lens' CreateGroupResponse (Maybe Text)

Latest version arn of the definition.

cgrsARN :: Lens' CreateGroupResponse (Maybe Text)

Arn of the definition.

cgrsName :: Lens' CreateGroupResponse (Maybe Text)

Name of the definition.

cgrsCreationTimestamp :: Lens' CreateGroupResponse (Maybe Text)

Timestamp of when the definition was created.

cgrsId :: Lens' CreateGroupResponse (Maybe Text)

Id of the definition.

cgrsLatestVersion :: Lens' CreateGroupResponse (Maybe Text)

Last version of the definition.

cgrsLastUpdatedTimestamp :: Lens' CreateGroupResponse (Maybe Text)

Last updated timestamp of the definition.

cgrsResponseStatus :: Lens' CreateGroupResponse Int

- | The response status code.