Network.AWS.Greengrass.CreateGroupCertificateAuthority

Description

Creates a CA for the group. If a CA already exists, it will rotate the existing CA.

Creating a Request

createGroupCertificateAuthority

Arguments

:: Text
cgcaGroupId

-> CreateGroupCertificateAuthority

Creates a value of CreateGroupCertificateAuthority with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgcaAmznClientToken - The client token used to request idempotent operations.

cgcaGroupId - The unique Id of the AWS Greengrass Group

data CreateGroupCertificateAuthority

See: createGroupCertificateAuthority smart constructor.

Request Lenses

cgcaAmznClientToken :: Lens' CreateGroupCertificateAuthority (Maybe Text)

The client token used to request idempotent operations.

cgcaGroupId :: Lens' CreateGroupCertificateAuthority Text

The unique Id of the AWS Greengrass Group

Destructuring the Response

createGroupCertificateAuthorityResponse

Arguments

:: Int
cgcarsResponseStatus

-> CreateGroupCertificateAuthorityResponse

Creates a value of CreateGroupCertificateAuthorityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgcarsGroupCertificateAuthorityARN - Arn of the group certificate authority.

cgcarsResponseStatus - -- | The response status code.

data CreateGroupCertificateAuthorityResponse

See: createGroupCertificateAuthorityResponse smart constructor.

Response Lenses

cgcarsGroupCertificateAuthorityARN :: Lens' CreateGroupCertificateAuthorityResponse (Maybe Text)

Arn of the group certificate authority.

cgcarsResponseStatus :: Lens' CreateGroupCertificateAuthorityResponse Int

- | The response status code.