Network.AWS.Greengrass.CreateGroupVersion

Description

Creates a version of a group which has already been defined.

Creating a Request

createGroupVersion

Arguments

:: Text
cgvGroupId

-> CreateGroupVersion

Creates a value of CreateGroupVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgvAmznClientToken - The client token used to request idempotent operations.

cgvSubscriptionDefinitionVersionARN - Subscription definition version arn for this group.

cgvCoreDefinitionVersionARN - Core definition version arn for this group.

cgvDeviceDefinitionVersionARN - Device definition version arn for this group.

cgvFunctionDefinitionVersionARN - Function definition version arn for this group.

cgvLoggerDefinitionVersionARN - Logger definitionv ersion arn for this group.

cgvGroupId - The unique Id of the AWS Greengrass Group

data CreateGroupVersion

See: createGroupVersion smart constructor.

Request Lenses

cgvAmznClientToken :: Lens' CreateGroupVersion (Maybe Text)

The client token used to request idempotent operations.

cgvSubscriptionDefinitionVersionARN :: Lens' CreateGroupVersion (Maybe Text)

Subscription definition version arn for this group.

cgvCoreDefinitionVersionARN :: Lens' CreateGroupVersion (Maybe Text)

Core definition version arn for this group.

cgvDeviceDefinitionVersionARN :: Lens' CreateGroupVersion (Maybe Text)

Device definition version arn for this group.

cgvFunctionDefinitionVersionARN :: Lens' CreateGroupVersion (Maybe Text)

Function definition version arn for this group.

cgvLoggerDefinitionVersionARN :: Lens' CreateGroupVersion (Maybe Text)

Logger definitionv ersion arn for this group.

cgvGroupId :: Lens' CreateGroupVersion Text

The unique Id of the AWS Greengrass Group

Destructuring the Response

createGroupVersionResponse

Arguments

:: Int
cgvrsResponseStatus

-> CreateGroupVersionResponse

Creates a value of CreateGroupVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgvrsARN - Arn of the version.

cgvrsCreationTimestamp - Timestamp of when the version was created.

cgvrsVersion - Unique Id of a version.

cgvrsId - Id of the resource container.

cgvrsResponseStatus - -- | The response status code.

data CreateGroupVersionResponse

See: createGroupVersionResponse smart constructor.

Response Lenses

cgvrsARN :: Lens' CreateGroupVersionResponse (Maybe Text)

Arn of the version.

cgvrsCreationTimestamp :: Lens' CreateGroupVersionResponse (Maybe Text)

Timestamp of when the version was created.

cgvrsVersion :: Lens' CreateGroupVersionResponse (Maybe Text)

Unique Id of a version.

cgvrsId :: Lens' CreateGroupVersionResponse (Maybe Text)

Id of the resource container.

cgvrsResponseStatus :: Lens' CreateGroupVersionResponse Int

- | The response status code.