Network.AWS.Greengrass.CreateLoggerDefinition

Description

Creates a logger definition. You may optionally provide the initial version of the logger definition or use `CreateLoggerDefinitionVersion` at a later time.

Creating a Request

createLoggerDefinition :: CreateLoggerDefinition

Creates a value of CreateLoggerDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cldAmznClientToken - The client token used to request idempotent operations.

cldInitialVersion - Information on the initial version

cldName - name of the logger definition

data CreateLoggerDefinition

See: createLoggerDefinition smart constructor.

Request Lenses

cldAmznClientToken :: Lens' CreateLoggerDefinition (Maybe Text)

The client token used to request idempotent operations.

cldInitialVersion :: Lens' CreateLoggerDefinition (Maybe LoggerDefinitionVersion)

Information on the initial version

cldName :: Lens' CreateLoggerDefinition (Maybe Text)

name of the logger definition

Destructuring the Response

createLoggerDefinitionResponse

Arguments

:: Int
cldrsResponseStatus

-> CreateLoggerDefinitionResponse

Creates a value of CreateLoggerDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cldrsLatestVersionARN - Latest version arn of the definition.

cldrsARN - Arn of the definition.

cldrsName - Name of the definition.

cldrsCreationTimestamp - Timestamp of when the definition was created.

cldrsId - Id of the definition.

cldrsLatestVersion - Last version of the definition.

cldrsLastUpdatedTimestamp - Last updated timestamp of the definition.

cldrsResponseStatus - -- | The response status code.

data CreateLoggerDefinitionResponse

See: createLoggerDefinitionResponse smart constructor.

Response Lenses

cldrsLatestVersionARN :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Latest version arn of the definition.

cldrsARN :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Arn of the definition.

cldrsName :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Name of the definition.

cldrsCreationTimestamp :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Timestamp of when the definition was created.

cldrsId :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Id of the definition.

cldrsLatestVersion :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Last version of the definition.

cldrsLastUpdatedTimestamp :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

Last updated timestamp of the definition.

cldrsResponseStatus :: Lens' CreateLoggerDefinitionResponse Int

- | The response status code.