Network.AWS.Greengrass.CreateSubscriptionDefinition

Description

Creates a subscription definition. You may optionally provide the initial version of the subscription definition or use `CreateSubscriptionDefinitionVersion` at a later time.

Creating a Request

createSubscriptionDefinition :: CreateSubscriptionDefinition

Creates a value of CreateSubscriptionDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csdAmznClientToken - The client token used to request idempotent operations.

csdInitialVersion - Information on the initial version

csdName - name of the subscription definition

data CreateSubscriptionDefinition

See: createSubscriptionDefinition smart constructor.

Request Lenses

csdAmznClientToken :: Lens' CreateSubscriptionDefinition (Maybe Text)

The client token used to request idempotent operations.

csdInitialVersion :: Lens' CreateSubscriptionDefinition (Maybe SubscriptionDefinitionVersion)

Information on the initial version

csdName :: Lens' CreateSubscriptionDefinition (Maybe Text)

name of the subscription definition

Destructuring the Response

createSubscriptionDefinitionResponse

Arguments

:: Int
csdrsResponseStatus

-> CreateSubscriptionDefinitionResponse

Creates a value of CreateSubscriptionDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csdrsLatestVersionARN - Latest version arn of the definition.

csdrsARN - Arn of the definition.

csdrsName - Name of the definition.

csdrsCreationTimestamp - Timestamp of when the definition was created.

csdrsId - Id of the definition.

csdrsLatestVersion - Last version of the definition.

csdrsLastUpdatedTimestamp - Last updated timestamp of the definition.

csdrsResponseStatus - -- | The response status code.

data CreateSubscriptionDefinitionResponse

See: createSubscriptionDefinitionResponse smart constructor.

Response Lenses

csdrsLatestVersionARN :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Latest version arn of the definition.

csdrsARN :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Arn of the definition.

csdrsName :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Name of the definition.

csdrsCreationTimestamp :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Timestamp of when the definition was created.

csdrsId :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Id of the definition.

csdrsLatestVersion :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Last version of the definition.

csdrsLastUpdatedTimestamp :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

Last updated timestamp of the definition.

csdrsResponseStatus :: Lens' CreateSubscriptionDefinitionResponse Int

- | The response status code.