Network.AWS.Greengrass.DeleteFunctionDefinition

Description

Deletes a Lambda function definition. The Lambda function definition must not have been used in a deployment.

Creating a Request

deleteFunctionDefinition

Arguments

:: Text
dfdFunctionDefinitionId

-> DeleteFunctionDefinition

Creates a value of DeleteFunctionDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfdFunctionDefinitionId - the unique Id of the lambda definition

data DeleteFunctionDefinition

See: deleteFunctionDefinition smart constructor.

Request Lenses

dfdFunctionDefinitionId :: Lens' DeleteFunctionDefinition Text

the unique Id of the lambda definition

Destructuring the Response

deleteFunctionDefinitionResponse

Arguments

:: Int
dfdrsResponseStatus

-> DeleteFunctionDefinitionResponse

Creates a value of DeleteFunctionDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfdrsResponseStatus - -- | The response status code.

data DeleteFunctionDefinitionResponse

See: deleteFunctionDefinitionResponse smart constructor.

Response Lenses

dfdrsResponseStatus :: Lens' DeleteFunctionDefinitionResponse Int

- | The response status code.