Network.AWS.Greengrass.DeleteGroup

Description

Deletes a group. The group must not have been used in deployment.

Creating a Request

deleteGroup

Arguments

:: Text
dgGroupId

-> DeleteGroup

Creates a value of DeleteGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgGroupId - The unique Id of the AWS Greengrass Group

data DeleteGroup

See: deleteGroup smart constructor.

Request Lenses

dgGroupId :: Lens' DeleteGroup Text

The unique Id of the AWS Greengrass Group

Destructuring the Response

deleteGroupResponse

Arguments

:: Int
dgrsResponseStatus

-> DeleteGroupResponse

Creates a value of DeleteGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgrsResponseStatus - -- | The response status code.

data DeleteGroupResponse

See: deleteGroupResponse smart constructor.

Response Lenses

dgrsResponseStatus :: Lens' DeleteGroupResponse Int

- | The response status code.