Network.AWS.Greengrass.GetAssociatedRole

Description

Retrieves the role associated with a particular group.

Creating a Request

getAssociatedRole

Arguments

:: Text
garGroupId

-> GetAssociatedRole

Creates a value of GetAssociatedRole with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garGroupId - The unique Id of the AWS Greengrass Group

data GetAssociatedRole

See: getAssociatedRole smart constructor.

Request Lenses

garGroupId :: Lens' GetAssociatedRole Text

The unique Id of the AWS Greengrass Group

Destructuring the Response

getAssociatedRoleResponse

Arguments

:: Int
garrsResponseStatus

-> GetAssociatedRoleResponse

Creates a value of GetAssociatedRoleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garrsAssociatedAt - Time when the role was associated for the group.

garrsRoleARN - Arn of the role that is associated with the group.

garrsResponseStatus - -- | The response status code.

data GetAssociatedRoleResponse

See: getAssociatedRoleResponse smart constructor.

Response Lenses

garrsAssociatedAt :: Lens' GetAssociatedRoleResponse (Maybe Text)

Time when the role was associated for the group.

garrsRoleARN :: Lens' GetAssociatedRoleResponse (Maybe Text)

Arn of the role that is associated with the group.

garrsResponseStatus :: Lens' GetAssociatedRoleResponse Int

- | The response status code.