Network.AWS.Greengrass.GetConnectivityInfo

Description

Retrieves the connectivity information for a core.

Creating a Request

getConnectivityInfo

Arguments

:: Text
gciThingName

-> GetConnectivityInfo

Creates a value of GetConnectivityInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gciThingName - Thing Name

data GetConnectivityInfo

See: getConnectivityInfo smart constructor.

Request Lenses

gciThingName :: Lens' GetConnectivityInfo Text

Thing Name

Destructuring the Response

getConnectivityInfoResponse

Arguments

:: Int
gcirsResponseStatus

-> GetConnectivityInfoResponse

Creates a value of GetConnectivityInfoResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcirsMessage - Response Text

gcirsConnectivityInfo - Connectivity info array

gcirsResponseStatus - -- | The response status code.

data GetConnectivityInfoResponse

See: getConnectivityInfoResponse smart constructor.

Response Lenses

gcirsMessage :: Lens' GetConnectivityInfoResponse (Maybe Text)

Response Text

gcirsConnectivityInfo :: Lens' GetConnectivityInfoResponse [ConnectivityInfo]

Connectivity info array

gcirsResponseStatus :: Lens' GetConnectivityInfoResponse Int

- | The response status code.