Network.AWS.Greengrass.GetCoreDefinition

Description

Retrieves information about a core definition version.

Creating a Request

getCoreDefinition

Arguments

:: Text
gcdCoreDefinitionId

-> GetCoreDefinition

Creates a value of GetCoreDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcdCoreDefinitionId - core definition Id

data GetCoreDefinition

See: getCoreDefinition smart constructor.

Request Lenses

gcdCoreDefinitionId :: Lens' GetCoreDefinition Text

core definition Id

Destructuring the Response

getCoreDefinitionResponse

Arguments

:: Int
gcdrsResponseStatus

-> GetCoreDefinitionResponse

Creates a value of GetCoreDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcdrsLatestVersionARN - Latest version arn of the definition.

gcdrsARN - Arn of the definition.

gcdrsName - Name of the definition.

gcdrsCreationTimestamp - Timestamp of when the definition was created.

gcdrsId - Id of the definition.

gcdrsLatestVersion - Last version of the definition.

gcdrsLastUpdatedTimestamp - Last updated timestamp of the definition.

gcdrsResponseStatus - -- | The response status code.

data GetCoreDefinitionResponse

See: getCoreDefinitionResponse smart constructor.

Response Lenses

gcdrsLatestVersionARN :: Lens' GetCoreDefinitionResponse (Maybe Text)

Latest version arn of the definition.

gcdrsARN :: Lens' GetCoreDefinitionResponse (Maybe Text)

Arn of the definition.

gcdrsName :: Lens' GetCoreDefinitionResponse (Maybe Text)

Name of the definition.

gcdrsCreationTimestamp :: Lens' GetCoreDefinitionResponse (Maybe Text)

Timestamp of when the definition was created.

gcdrsId :: Lens' GetCoreDefinitionResponse (Maybe Text)

Id of the definition.

gcdrsLatestVersion :: Lens' GetCoreDefinitionResponse (Maybe Text)

Last version of the definition.

gcdrsLastUpdatedTimestamp :: Lens' GetCoreDefinitionResponse (Maybe Text)

Last updated timestamp of the definition.

gcdrsResponseStatus :: Lens' GetCoreDefinitionResponse Int

- | The response status code.