Network.AWS.Greengrass.GetDeploymentStatus

Description

Returns the status of a deployment.

Creating a Request

getDeploymentStatus

Arguments

:: Text
gdsGroupId

-> Text
gdsDeploymentId

-> GetDeploymentStatus

Creates a value of GetDeploymentStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsGroupId - The unique Id of the AWS Greengrass Group

gdsDeploymentId - the deployment Id

data GetDeploymentStatus

See: getDeploymentStatus smart constructor.

Request Lenses

gdsGroupId :: Lens' GetDeploymentStatus Text

The unique Id of the AWS Greengrass Group

gdsDeploymentId :: Lens' GetDeploymentStatus Text

the deployment Id

Destructuring the Response

getDeploymentStatusResponse

Arguments

:: Int
gdsrsResponseStatus

-> GetDeploymentStatusResponse

Creates a value of GetDeploymentStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsDeploymentStatus - Status of the deployment.

gdsrsUpdatedAt - Last time the deployment status was updated.

gdsrsErrorMessage - Error Message

gdsrsResponseStatus - -- | The response status code.

data GetDeploymentStatusResponse

See: getDeploymentStatusResponse smart constructor.

Response Lenses

gdsrsDeploymentStatus :: Lens' GetDeploymentStatusResponse (Maybe Text)

Status of the deployment.

gdsrsUpdatedAt :: Lens' GetDeploymentStatusResponse (Maybe Text)

Last time the deployment status was updated.

gdsrsErrorMessage :: Lens' GetDeploymentStatusResponse (Maybe Text)

Error Message

gdsrsResponseStatus :: Lens' GetDeploymentStatusResponse Int

- | The response status code.