Network.AWS.Greengrass.GetDeviceDefinitionVersion

Description

Retrieves information about a device definition version.

Creating a Request

getDeviceDefinitionVersion

Arguments

:: Text
gddvDeviceDefinitionVersionId

-> Text
gddvDeviceDefinitionId

-> GetDeviceDefinitionVersion

Creates a value of GetDeviceDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gddvDeviceDefinitionVersionId - device definition version Id

gddvDeviceDefinitionId - device definition Id

data GetDeviceDefinitionVersion

See: getDeviceDefinitionVersion smart constructor.

Request Lenses

gddvDeviceDefinitionVersionId :: Lens' GetDeviceDefinitionVersion Text

device definition version Id

gddvDeviceDefinitionId :: Lens' GetDeviceDefinitionVersion Text

device definition Id

Destructuring the Response

getDeviceDefinitionVersionResponse

Arguments

:: Int
gddvrsResponseStatus

-> GetDeviceDefinitionVersionResponse

Creates a value of GetDeviceDefinitionVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gddvrsDefinition - Device definition version

gddvrsARN - Arn of the device definition version.

gddvrsCreationTimestamp - Timestamp of when the device definition version was created.

gddvrsVersion - Version of the device definition version.

gddvrsId - Id of the device definition the version belongs to.

gddvrsResponseStatus - -- | The response status code.

data GetDeviceDefinitionVersionResponse

See: getDeviceDefinitionVersionResponse smart constructor.

Response Lenses

gddvrsDefinition :: Lens' GetDeviceDefinitionVersionResponse (Maybe DeviceDefinitionVersion)

Device definition version

gddvrsARN :: Lens' GetDeviceDefinitionVersionResponse (Maybe Text)

Arn of the device definition version.

gddvrsCreationTimestamp :: Lens' GetDeviceDefinitionVersionResponse (Maybe Text)

Timestamp of when the device definition version was created.

gddvrsVersion :: Lens' GetDeviceDefinitionVersionResponse (Maybe Text)

Version of the device definition version.

gddvrsId :: Lens' GetDeviceDefinitionVersionResponse (Maybe Text)

Id of the device definition the version belongs to.

gddvrsResponseStatus :: Lens' GetDeviceDefinitionVersionResponse Int

- | The response status code.