Network.AWS.Greengrass.GetFunctionDefinition

Description

Retrieves information about a Lambda function definition, such as its creation time and latest version.

Creating a Request

getFunctionDefinition

Arguments

:: Text
gfdFunctionDefinitionId

-> GetFunctionDefinition

Creates a value of GetFunctionDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gfdFunctionDefinitionId - the unique Id of the lambda definition

data GetFunctionDefinition

See: getFunctionDefinition smart constructor.

Request Lenses

gfdFunctionDefinitionId :: Lens' GetFunctionDefinition Text

the unique Id of the lambda definition

Destructuring the Response

getFunctionDefinitionResponse

Arguments

:: Int
gfdrsResponseStatus

-> GetFunctionDefinitionResponse

Creates a value of GetFunctionDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gfdrsLatestVersionARN - Latest version arn of the definition.

gfdrsARN - Arn of the definition.

gfdrsName - Name of the definition.

gfdrsCreationTimestamp - Timestamp of when the definition was created.

gfdrsId - Id of the definition.

gfdrsLatestVersion - Last version of the definition.

gfdrsLastUpdatedTimestamp - Last updated timestamp of the definition.

gfdrsResponseStatus - -- | The response status code.

data GetFunctionDefinitionResponse

See: getFunctionDefinitionResponse smart constructor.

Response Lenses

gfdrsLatestVersionARN :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Latest version arn of the definition.

gfdrsARN :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Arn of the definition.

gfdrsName :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Name of the definition.

gfdrsCreationTimestamp :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Timestamp of when the definition was created.

gfdrsId :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Id of the definition.

gfdrsLatestVersion :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Last version of the definition.

gfdrsLastUpdatedTimestamp :: Lens' GetFunctionDefinitionResponse (Maybe Text)

Last updated timestamp of the definition.

gfdrsResponseStatus :: Lens' GetFunctionDefinitionResponse Int

- | The response status code.