Network.AWS.Greengrass.GetGroup

Description

Retrieves information about a group.

Creating a Request

getGroup

Arguments

:: Text
ggGroupId

-> GetGroup

Creates a value of GetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ggGroupId - The unique Id of the AWS Greengrass Group

data GetGroup

See: getGroup smart constructor.

Request Lenses

ggGroupId :: Lens' GetGroup Text

The unique Id of the AWS Greengrass Group

Destructuring the Response

getGroupResponse

Arguments

:: Int
ggrsResponseStatus

-> GetGroupResponse

Creates a value of GetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ggrsLatestVersionARN - Latest version arn of the definition.

ggrsARN - Arn of the definition.

ggrsName - Name of the definition.

ggrsCreationTimestamp - Timestamp of when the definition was created.

ggrsId - Id of the definition.

ggrsLatestVersion - Last version of the definition.

ggrsLastUpdatedTimestamp - Last updated timestamp of the definition.

ggrsResponseStatus - -- | The response status code.

data GetGroupResponse

See: getGroupResponse smart constructor.

Response Lenses

ggrsLatestVersionARN :: Lens' GetGroupResponse (Maybe Text)

Latest version arn of the definition.

ggrsARN :: Lens' GetGroupResponse (Maybe Text)

Arn of the definition.

ggrsName :: Lens' GetGroupResponse (Maybe Text)

Name of the definition.

ggrsCreationTimestamp :: Lens' GetGroupResponse (Maybe Text)

Timestamp of when the definition was created.

ggrsId :: Lens' GetGroupResponse (Maybe Text)

Id of the definition.

ggrsLatestVersion :: Lens' GetGroupResponse (Maybe Text)

Last version of the definition.

ggrsLastUpdatedTimestamp :: Lens' GetGroupResponse (Maybe Text)

Last updated timestamp of the definition.

ggrsResponseStatus :: Lens' GetGroupResponse Int

- | The response status code.