Network.AWS.Greengrass.GetLoggerDefinitionVersion

Description

Retrieves information about a logger definition version.

Creating a Request

getLoggerDefinitionVersion

Arguments

:: Text
gldvLoggerDefinitionVersionId

-> Text
gldvLoggerDefinitionId

-> GetLoggerDefinitionVersion

Creates a value of GetLoggerDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gldvLoggerDefinitionVersionId - logger definition version Id

gldvLoggerDefinitionId - logger definition Id

data GetLoggerDefinitionVersion

See: getLoggerDefinitionVersion smart constructor.

Request Lenses

gldvLoggerDefinitionVersionId :: Lens' GetLoggerDefinitionVersion Text

logger definition version Id

gldvLoggerDefinitionId :: Lens' GetLoggerDefinitionVersion Text

logger definition Id

Destructuring the Response

getLoggerDefinitionVersionResponse

Arguments

:: Int
gldvrsResponseStatus

-> GetLoggerDefinitionVersionResponse

Creates a value of GetLoggerDefinitionVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gldvrsDefinition - Information on definition

gldvrsARN - Arn of the logger definition version.

gldvrsCreationTimestamp - Timestamp of when the logger definition version was created.

gldvrsVersion - Version of the logger definition version.

gldvrsId - Id of the logger definition the version belongs to.

gldvrsResponseStatus - -- | The response status code.

data GetLoggerDefinitionVersionResponse

See: getLoggerDefinitionVersionResponse smart constructor.

Response Lenses

gldvrsDefinition :: Lens' GetLoggerDefinitionVersionResponse (Maybe LoggerDefinitionVersion)

Information on definition

gldvrsARN :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

Arn of the logger definition version.

gldvrsCreationTimestamp :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

Timestamp of when the logger definition version was created.

gldvrsVersion :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

Version of the logger definition version.

gldvrsId :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

Id of the logger definition the version belongs to.

gldvrsResponseStatus :: Lens' GetLoggerDefinitionVersionResponse Int

- | The response status code.