Network.AWS.Greengrass.GetSubscriptionDefinition

Description

Retrieves information about a subscription definition.

Creating a Request

getSubscriptionDefinition

Arguments

:: Text
gsdSubscriptionDefinitionId

-> GetSubscriptionDefinition

Creates a value of GetSubscriptionDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdSubscriptionDefinitionId - subscription definition Id

data GetSubscriptionDefinition

See: getSubscriptionDefinition smart constructor.

Request Lenses

gsdSubscriptionDefinitionId :: Lens' GetSubscriptionDefinition Text

subscription definition Id

Destructuring the Response

getSubscriptionDefinitionResponse

Arguments

:: Int
gsdrsResponseStatus

-> GetSubscriptionDefinitionResponse

Creates a value of GetSubscriptionDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdrsLatestVersionARN - Latest version arn of the definition.

gsdrsARN - Arn of the definition.

gsdrsName - Name of the definition.

gsdrsCreationTimestamp - Timestamp of when the definition was created.

gsdrsId - Id of the definition.

gsdrsLatestVersion - Last version of the definition.

gsdrsLastUpdatedTimestamp - Last updated timestamp of the definition.

gsdrsResponseStatus - -- | The response status code.

data GetSubscriptionDefinitionResponse

See: getSubscriptionDefinitionResponse smart constructor.

Response Lenses

gsdrsLatestVersionARN :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Latest version arn of the definition.

gsdrsARN :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Arn of the definition.

gsdrsName :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Name of the definition.

gsdrsCreationTimestamp :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Timestamp of when the definition was created.

gsdrsId :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Id of the definition.

gsdrsLatestVersion :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Last version of the definition.

gsdrsLastUpdatedTimestamp :: Lens' GetSubscriptionDefinitionResponse (Maybe Text)

Last updated timestamp of the definition.

gsdrsResponseStatus :: Lens' GetSubscriptionDefinitionResponse Int

- | The response status code.