Network.AWS.Greengrass.GetSubscriptionDefinitionVersion

Description

Retrieves information about a subscription definition version.

Creating a Request

getSubscriptionDefinitionVersion

Arguments

:: Text
gsdvSubscriptionDefinitionId

-> Text
gsdvSubscriptionDefinitionVersionId

-> GetSubscriptionDefinitionVersion

Creates a value of GetSubscriptionDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdvSubscriptionDefinitionId - subscription definition Id

gsdvSubscriptionDefinitionVersionId - subscription definition version Id

data GetSubscriptionDefinitionVersion

See: getSubscriptionDefinitionVersion smart constructor.

Request Lenses

gsdvSubscriptionDefinitionId :: Lens' GetSubscriptionDefinitionVersion Text

subscription definition Id

gsdvSubscriptionDefinitionVersionId :: Lens' GetSubscriptionDefinitionVersion Text

subscription definition version Id

Destructuring the Response

getSubscriptionDefinitionVersionResponse

Arguments

:: Int
gsdvrsResponseStatus

-> GetSubscriptionDefinitionVersionResponse

Creates a value of GetSubscriptionDefinitionVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdvrsDefinition - Information on the definition

gsdvrsARN - Arn of the subscription definition version.

gsdvrsCreationTimestamp - Timestamp of when the subscription definition version was created.

gsdvrsVersion - Version of the subscription definition version.

gsdvrsId - Id of the subscription definition the version belongs to.

gsdvrsResponseStatus - -- | The response status code.

data GetSubscriptionDefinitionVersionResponse

See: getSubscriptionDefinitionVersionResponse smart constructor.

Response Lenses

gsdvrsDefinition :: Lens' GetSubscriptionDefinitionVersionResponse (Maybe SubscriptionDefinitionVersion)

Information on the definition

gsdvrsARN :: Lens' GetSubscriptionDefinitionVersionResponse (Maybe Text)

Arn of the subscription definition version.

gsdvrsCreationTimestamp :: Lens' GetSubscriptionDefinitionVersionResponse (Maybe Text)

Timestamp of when the subscription definition version was created.

gsdvrsVersion :: Lens' GetSubscriptionDefinitionVersionResponse (Maybe Text)

Version of the subscription definition version.

gsdvrsId :: Lens' GetSubscriptionDefinitionVersionResponse (Maybe Text)

Id of the subscription definition the version belongs to.

gsdvrsResponseStatus :: Lens' GetSubscriptionDefinitionVersionResponse Int

- | The response status code.