Network.AWS.Greengrass.ListCoreDefinitions

Description

Retrieves a list of core definitions.

Creating a Request

listCoreDefinitions :: ListCoreDefinitions

Creates a value of ListCoreDefinitions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcdNextToken - Specifies the pagination token used when iterating through a paginated request

lcdMaxResults - Specifies the maximum number of list results to be returned in this page

data ListCoreDefinitions

See: listCoreDefinitions smart constructor.

Request Lenses

lcdNextToken :: Lens' ListCoreDefinitions (Maybe Text)

Specifies the pagination token used when iterating through a paginated request

lcdMaxResults :: Lens' ListCoreDefinitions (Maybe Text)

Specifies the maximum number of list results to be returned in this page

Destructuring the Response

listCoreDefinitionsResponse

Arguments

:: Int
lcdrsResponseStatus

-> ListCoreDefinitionsResponse

Creates a value of ListCoreDefinitionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcdrsNextToken - The token for the next set of results, or 'null' if there are no additional results.

lcdrsDefinitions - Definitions

lcdrsResponseStatus - -- | The response status code.

data ListCoreDefinitionsResponse

See: listCoreDefinitionsResponse smart constructor.

Response Lenses

lcdrsNextToken :: Lens' ListCoreDefinitionsResponse (Maybe Text)

The token for the next set of results, or 'null' if there are no additional results.

lcdrsDefinitions :: Lens' ListCoreDefinitionsResponse [DefinitionInformation]

Definitions

lcdrsResponseStatus :: Lens' ListCoreDefinitionsResponse Int

- | The response status code.