Network.AWS.Greengrass.ListLoggerDefinitionVersions

Description

Lists the versions of a logger definition.

Creating a Request

listLoggerDefinitionVersions

Arguments

:: Text
lldvLoggerDefinitionId

-> ListLoggerDefinitionVersions

Creates a value of ListLoggerDefinitionVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lldvNextToken - Specifies the pagination token used when iterating through a paginated request

lldvMaxResults - Specifies the maximum number of list results to be returned in this page

lldvLoggerDefinitionId - logger definition Id

data ListLoggerDefinitionVersions

See: listLoggerDefinitionVersions smart constructor.

Request Lenses

lldvNextToken :: Lens' ListLoggerDefinitionVersions (Maybe Text)

Specifies the pagination token used when iterating through a paginated request

lldvMaxResults :: Lens' ListLoggerDefinitionVersions (Maybe Text)

Specifies the maximum number of list results to be returned in this page

lldvLoggerDefinitionId :: Lens' ListLoggerDefinitionVersions Text

logger definition Id

Destructuring the Response

listLoggerDefinitionVersionsResponse

Arguments

:: Int
lldvrsResponseStatus

-> ListLoggerDefinitionVersionsResponse

Creates a value of ListLoggerDefinitionVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lldvrsVersions - Versions

lldvrsNextToken - The token for the next set of results, or 'null' if there are no additional results.

lldvrsResponseStatus - -- | The response status code.

data ListLoggerDefinitionVersionsResponse

See: listLoggerDefinitionVersionsResponse smart constructor.

Response Lenses

lldvrsVersions :: Lens' ListLoggerDefinitionVersionsResponse [VersionInformation]

Versions

lldvrsNextToken :: Lens' ListLoggerDefinitionVersionsResponse (Maybe Text)

The token for the next set of results, or 'null' if there are no additional results.

lldvrsResponseStatus :: Lens' ListLoggerDefinitionVersionsResponse Int

- | The response status code.