Network.AWS.Greengrass.Types

Description

Service Configuration

greengrass :: Service

API version 2017-06-07 of the Amazon Greengrass SDK configuration.

Errors

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

General Error

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

General Error

DeploymentType

data DeploymentType

Constructors

NewDeployment

Redeployment

LoggerComponent

data LoggerComponent

Constructors

GreengrassSystem

Lambda

LoggerLevel

data LoggerLevel

Constructors

Debug

Error'

Fatal

Info

Warn

LoggerType

data LoggerType

Constructors

AWSCloudWatch

FileSystem

ConnectivityInfo

data ConnectivityInfo

Connectivity Info

See: connectivityInfo smart constructor.

connectivityInfo :: ConnectivityInfo

Creates a value of ConnectivityInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciPortNumber - Port of the GGC. Usually 8883.

ciId - Element Id for this entry in the list.

ciMetadata - Metadata for this endpoint.

ciHostAddress - Endpoint for the GGC. Can be an IP address or DNS.

ciPortNumber :: Lens' ConnectivityInfo (Maybe Int)

Port of the GGC. Usually 8883.

ciId :: Lens' ConnectivityInfo (Maybe Text)

Element Id for this entry in the list.

ciMetadata :: Lens' ConnectivityInfo (Maybe Text)

Metadata for this endpoint.

ciHostAddress :: Lens' ConnectivityInfo (Maybe Text)

Endpoint for the GGC. Can be an IP address or DNS.

Core

data Core

Information on the core

See: core smart constructor.

core :: Core

Creates a value of Core with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cCertificateARN - Certificate arn of the core.

cThingARN - Thing arn of the core.

cSyncShadow - If true, the local shadow value automatically syncs with the cloud's shadow state.

cId - Element Id for this entry in the list.

cCertificateARN :: Lens' Core (Maybe Text)

Certificate arn of the core.

cThingARN :: Lens' Core (Maybe Text)

Thing arn of the core.

cSyncShadow :: Lens' Core (Maybe Bool)

If true, the local shadow value automatically syncs with the cloud's shadow state.

cId :: Lens' Core (Maybe Text)

Element Id for this entry in the list.

CoreDefinitionVersion

data CoreDefinitionVersion

Information on core definition version

See: coreDefinitionVersion smart constructor.

coreDefinitionVersion :: CoreDefinitionVersion

Creates a value of CoreDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdvCores - Cores in the definition version.

cdvCores :: Lens' CoreDefinitionVersion [Core]

Cores in the definition version.

DefinitionInformation

data DefinitionInformation

Information on the Definition

See: definitionInformation smart constructor.

definitionInformation :: DefinitionInformation

Creates a value of DefinitionInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diLatestVersionARN - Latest version arn of the definition.

diARN - Arn of the definition.

diName - Name of the definition.

diCreationTimestamp - Timestamp of when the definition was created.

diId - Id of the definition.

diLatestVersion - Last version of the definition.

diLastUpdatedTimestamp - Last updated timestamp of the definition.

diLatestVersionARN :: Lens' DefinitionInformation (Maybe Text)

Latest version arn of the definition.

diARN :: Lens' DefinitionInformation (Maybe Text)

Arn of the definition.

diName :: Lens' DefinitionInformation (Maybe Text)

Name of the definition.

diCreationTimestamp :: Lens' DefinitionInformation (Maybe Text)

Timestamp of when the definition was created.

diId :: Lens' DefinitionInformation (Maybe Text)

Id of the definition.

diLatestVersion :: Lens' DefinitionInformation (Maybe Text)

Last version of the definition.

diLastUpdatedTimestamp :: Lens' DefinitionInformation (Maybe Text)

Last updated timestamp of the definition.

Deployment

data Deployment

Information on the deployment

See: deployment smart constructor.

deployment :: Deployment

Creates a value of Deployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dDeploymentId - Id of the deployment.

dDeploymentARN - Arn of the deployment.

dCreatedAt - Timestamp when the deployment was created.

dGroupARN - Arn of the group for this deployment.

dDeploymentId :: Lens' Deployment (Maybe Text)

Id of the deployment.

dDeploymentARN :: Lens' Deployment (Maybe Text)

Arn of the deployment.

dCreatedAt :: Lens' Deployment (Maybe Text)

Timestamp when the deployment was created.

dGroupARN :: Lens' Deployment (Maybe Text)

Arn of the group for this deployment.

Device

data Device

Information on a Device

See: device smart constructor.

device :: Device

Creates a value of Device with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dCertificateARN - Certificate arn of the device.

dThingARN - Thing arn of the device.

dSyncShadow - If true, the local shadow value automatically syncs with the cloud's shadow state.

dId - Element Id for this entry in the list.

dCertificateARN :: Lens' Device (Maybe Text)

Certificate arn of the device.

dThingARN :: Lens' Device (Maybe Text)

Thing arn of the device.

dSyncShadow :: Lens' Device (Maybe Bool)

If true, the local shadow value automatically syncs with the cloud's shadow state.

dId :: Lens' Device (Maybe Text)

Element Id for this entry in the list.

DeviceDefinitionVersion

data DeviceDefinitionVersion

Information on device definition version

See: deviceDefinitionVersion smart constructor.

deviceDefinitionVersion :: DeviceDefinitionVersion

Creates a value of DeviceDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddvDevices - Devices in the definition version.

ddvDevices :: Lens' DeviceDefinitionVersion [Device]

Devices in the definition version.

Function

data Function

Information on function

See: function smart constructor.

function :: Function

Creates a value of Function with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fFunctionARN - Arn of the Lambda function.

fFunctionConfiguration - Configuration of the function

fId - Id of the function in this version.

fFunctionARN :: Lens' Function (Maybe Text)

Arn of the Lambda function.

fFunctionConfiguration :: Lens' Function (Maybe FunctionConfiguration)

Configuration of the function

fId :: Lens' Function (Maybe Text)

Id of the function in this version.

FunctionConfiguration

data FunctionConfiguration

Configuration of the function

See: functionConfiguration smart constructor.

functionConfiguration :: FunctionConfiguration

Creates a value of FunctionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fcMemorySize - The memory size, in KB, you configured for the function.

fcExecArgs - Execution Arguments

fcEnvironment - Environment of the function configuration

fcExecutable - Executable

fcPinned - Whether the function is pinned or not. Pinned means the function is long-lived and starts when the core starts.

fcTimeout - The function execution time at which Lambda should terminate the function. This timeout still applies to pinned lambdas for each request.

fcMemorySize :: Lens' FunctionConfiguration (Maybe Int)

The memory size, in KB, you configured for the function.

fcExecArgs :: Lens' FunctionConfiguration (Maybe Text)

Execution Arguments

fcEnvironment :: Lens' FunctionConfiguration (Maybe FunctionConfigurationEnvironment)

Environment of the function configuration

fcExecutable :: Lens' FunctionConfiguration (Maybe Text)

Executable

fcPinned :: Lens' FunctionConfiguration (Maybe Bool)

Whether the function is pinned or not. Pinned means the function is long-lived and starts when the core starts.

fcTimeout :: Lens' FunctionConfiguration (Maybe Int)

The function execution time at which Lambda should terminate the function. This timeout still applies to pinned lambdas for each request.

FunctionConfigurationEnvironment

data FunctionConfigurationEnvironment

Environment of the function configuration

See: functionConfigurationEnvironment smart constructor.

functionConfigurationEnvironment :: FunctionConfigurationEnvironment

Creates a value of FunctionConfigurationEnvironment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fceVariables - Environment variables for the lambda function.

fceVariables :: Lens' FunctionConfigurationEnvironment (HashMap Text Text)

Environment variables for the lambda function.

FunctionDefinitionVersion

data FunctionDefinitionVersion

Information on the function definition version

See: functionDefinitionVersion smart constructor.

functionDefinitionVersion :: FunctionDefinitionVersion

Creates a value of FunctionDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fdvFunctions - Lambda functions in this function definition version.

fdvFunctions :: Lens' FunctionDefinitionVersion [Function]

Lambda functions in this function definition version.

GreengrassLogger

data GreengrassLogger

Information on the Logger

See: greengrassLogger smart constructor.

greengrassLogger :: GreengrassLogger

Creates a value of GreengrassLogger with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

glSpace - Amount of hardware space, in KB, to use if file system is used for logging purposes.

glComponent - The component that will be subject to logs

glId - Element Id for this entry in the list.

glType - The type which will be use for log output

glLevel - The level of the logs

glSpace :: Lens' GreengrassLogger (Maybe Int)

Amount of hardware space, in KB, to use if file system is used for logging purposes.

glComponent :: Lens' GreengrassLogger (Maybe LoggerComponent)

The component that will be subject to logs

glId :: Lens' GreengrassLogger (Maybe Text)

Element Id for this entry in the list.

glType :: Lens' GreengrassLogger (Maybe LoggerType)

The type which will be use for log output

glLevel :: Lens' GreengrassLogger (Maybe LoggerLevel)

The level of the logs

GroupCertificateAuthorityProperties

data GroupCertificateAuthorityProperties

Information on group certificate authority properties

See: groupCertificateAuthorityProperties smart constructor.

groupCertificateAuthorityProperties :: GroupCertificateAuthorityProperties

Creates a value of GroupCertificateAuthorityProperties with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcapGroupCertificateAuthorityARN - Arn of the certificate authority for the group.

gcapGroupCertificateAuthorityId - Id of the certificate authority for the group.

gcapGroupCertificateAuthorityARN :: Lens' GroupCertificateAuthorityProperties (Maybe Text)

Arn of the certificate authority for the group.

gcapGroupCertificateAuthorityId :: Lens' GroupCertificateAuthorityProperties (Maybe Text)

Id of the certificate authority for the group.

GroupInformation

data GroupInformation

Information of a group

See: groupInformation smart constructor.

groupInformation :: GroupInformation

Creates a value of GroupInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giLatestVersionARN - Latest version arn of the group.

giARN - Arn of a group.

giName - Name of a group.

giCreationTimestamp - Timestamp of when the group was created.

giId - Id of a group.

giLatestVersion - Last version of the group.

giLastUpdatedTimestamp - Last updated timestamp of the group.

giLatestVersionARN :: Lens' GroupInformation (Maybe Text)

Latest version arn of the group.

giARN :: Lens' GroupInformation (Maybe Text)

Arn of a group.

giName :: Lens' GroupInformation (Maybe Text)

Name of a group.

giCreationTimestamp :: Lens' GroupInformation (Maybe Text)

Timestamp of when the group was created.

giId :: Lens' GroupInformation (Maybe Text)

Id of a group.

giLatestVersion :: Lens' GroupInformation (Maybe Text)

Last version of the group.

giLastUpdatedTimestamp :: Lens' GroupInformation (Maybe Text)

Last updated timestamp of the group.

GroupVersion

data GroupVersion

Information on group version

See: groupVersion smart constructor.

groupVersion :: GroupVersion

Creates a value of GroupVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gvSubscriptionDefinitionVersionARN - Subscription definition version arn for this group.

gvCoreDefinitionVersionARN - Core definition version arn for this group.

gvDeviceDefinitionVersionARN - Device definition version arn for this group.

gvFunctionDefinitionVersionARN - Function definition version arn for this group.

gvLoggerDefinitionVersionARN - Logger definitionv ersion arn for this group.

gvSubscriptionDefinitionVersionARN :: Lens' GroupVersion (Maybe Text)

Subscription definition version arn for this group.

gvCoreDefinitionVersionARN :: Lens' GroupVersion (Maybe Text)

Core definition version arn for this group.

gvDeviceDefinitionVersionARN :: Lens' GroupVersion (Maybe Text)

Device definition version arn for this group.

gvFunctionDefinitionVersionARN :: Lens' GroupVersion (Maybe Text)

Function definition version arn for this group.

gvLoggerDefinitionVersionARN :: Lens' GroupVersion (Maybe Text)

Logger definitionv ersion arn for this group.

LoggerDefinitionVersion

data LoggerDefinitionVersion

Information on logger definition version

See: loggerDefinitionVersion smart constructor.

loggerDefinitionVersion :: LoggerDefinitionVersion

Creates a value of LoggerDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldvLoggers - List of loggers.

ldvLoggers :: Lens' LoggerDefinitionVersion [GreengrassLogger]

List of loggers.

Subscription

data Subscription

Information on subscription

See: subscription smart constructor.

subscription :: Subscription

Creates a value of Subscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sSubject - Subject of the message.

sSource - Source of the subscription. Can be a thing arn, lambda arn or word cloud

sId - Element Id for this entry in the list.

sTarget - Where the message is sent to. Can be a thing arn, lambda arn or word cloud.

sSubject :: Lens' Subscription (Maybe Text)

Subject of the message.

sSource :: Lens' Subscription (Maybe Text)

Source of the subscription. Can be a thing arn, lambda arn or word cloud

sId :: Lens' Subscription (Maybe Text)

Element Id for this entry in the list.

sTarget :: Lens' Subscription (Maybe Text)

Where the message is sent to. Can be a thing arn, lambda arn or word cloud.

SubscriptionDefinitionVersion

data SubscriptionDefinitionVersion

Information on subscription definition version

See: subscriptionDefinitionVersion smart constructor.

subscriptionDefinitionVersion :: SubscriptionDefinitionVersion

Creates a value of SubscriptionDefinitionVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdvSubscriptions - Subscriptions in the version.

sdvSubscriptions :: Lens' SubscriptionDefinitionVersion [Subscription]

Subscriptions in the version.

VersionInformation

data VersionInformation

Information on the version

See: versionInformation smart constructor.

versionInformation :: VersionInformation

Creates a value of VersionInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

viARN - Arn of the version.

viCreationTimestamp - Timestamp of when the version was created.

viVersion - Unique Id of a version.

viId - Id of the resource container.

viARN :: Lens' VersionInformation (Maybe Text)

Arn of the version.

viCreationTimestamp :: Lens' VersionInformation (Maybe Text)

Timestamp of when the version was created.

viVersion :: Lens' VersionInformation (Maybe Text)

Unique Id of a version.

viId :: Lens' VersionInformation (Maybe Text)

Id of the resource container.