Network.AWS.Greengrass.UpdateGroupCertificateConfiguration

Description

Updates the Cert expiry time for a group.

Creating a Request

updateGroupCertificateConfiguration

Arguments

:: Text
ugccGroupId

-> UpdateGroupCertificateConfiguration

Creates a value of UpdateGroupCertificateConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugccCertificateExpiryInMilliseconds - Amount of time when the certificate expires in milliseconds.

ugccGroupId - The unique Id of the AWS Greengrass Group

data UpdateGroupCertificateConfiguration

See: updateGroupCertificateConfiguration smart constructor.

Request Lenses

ugccCertificateExpiryInMilliseconds :: Lens' UpdateGroupCertificateConfiguration (Maybe Text)

Amount of time when the certificate expires in milliseconds.

ugccGroupId :: Lens' UpdateGroupCertificateConfiguration Text

The unique Id of the AWS Greengrass Group

Destructuring the Response

updateGroupCertificateConfigurationResponse

Arguments

:: Int
ugccrsResponseStatus

-> UpdateGroupCertificateConfigurationResponse

Creates a value of UpdateGroupCertificateConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugccrsCertificateAuthorityExpiryInMilliseconds - Amount of time when the certificate authority expires in milliseconds.

ugccrsGroupId - Id of the group the certificate configuration belongs to.

ugccrsCertificateExpiryInMilliseconds - Amount of time when the certificate expires in milliseconds.

ugccrsResponseStatus - -- | The response status code.

data UpdateGroupCertificateConfigurationResponse

See: updateGroupCertificateConfigurationResponse smart constructor.

Response Lenses

ugccrsCertificateAuthorityExpiryInMilliseconds :: Lens' UpdateGroupCertificateConfigurationResponse (Maybe Text)

Amount of time when the certificate authority expires in milliseconds.

ugccrsGroupId :: Lens' UpdateGroupCertificateConfigurationResponse (Maybe Text)

Id of the group the certificate configuration belongs to.

ugccrsCertificateExpiryInMilliseconds :: Lens' UpdateGroupCertificateConfigurationResponse (Maybe Text)

Amount of time when the certificate expires in milliseconds.

ugccrsResponseStatus :: Lens' UpdateGroupCertificateConfigurationResponse Int

- | The response status code.